Auth & Middleware

+++ routes/web.php
@@ -20,34 +20,34 @@ Route::get('/', [ListingController::class, 'index']);
 
 
 //Create Post
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 //Single Post
-Route::POST('/listings', [ListingController::class, 'store']);
+Route::POST('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 //Edit Post
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 //Update Post 
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 //Delete Post
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 //Single Post
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 //Show Register/Create Form
-Route::get('/register',[UserController::class, 'create']);
+Route::get('/register',[UserController::class, 'create'])->middleware('guest');
 
 //Create new User
 Route::post('/users',[UserController::class, 'store']);
 
 //Log User Out
-Route::post('/logout',[UserController::class, 'logout']); 
+Route::post('/logout',[UserController::class, 'logout'])->middleware('auth'); 
 
 //Show Log In
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 // Log In User
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
